Implement UserSchema using string concatenation
